Desktop FPS game summary+additions.
For the Player Character script I created a system to check for when the health of the player reaches 0 and to return a boolean variable of whether or not they are still alive or not.
For the FPS Input script I added a jump mechanic that only allows you to jump when you are standing on the ground. I also implemented the dead or alive tracker to the characters movements so if the player is alive they will be able to move and jump but if they are dead they won't be able to move at all.
For the RayShooter script I implemented the dead or alive tracker so that when you die you can't shoot anymore.
(Side note I did have to checkmark the notDead box for PlayerCharacter under the Player game object in order to run the game at first(May not be required after downloading files))
I also added a Borg Skin to the Cubes.